Cleaned up controls, favorite toggling and helper structure
