+++ QuickStackStore/QuickStackStore.csproj
@@ -90,6 +90,8 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QuickStackStorePlugin.cs" />
+    <Compile Include="Source\Controls\FavoritingMode.cs" />
+    <Compile Include="Source\Controls\HotkeyListener.cs" />
     <Compile Include="Source\LocalizationConfig.cs" />
     <Compile Include="Source\Modules\CompatibilitySupport.cs" />
     <Compile Include="Source\ContainerFinder.cs" />
