+++ QuickStackStore/Source/Helper.cs
@@ -1,5 +1,4 @@
-﻿using HarmonyLib;
-using System.IO;
+﻿using System.IO;
 using System.Reflection;
 using UnityEngine;
 using static QuickStackStore.QSSConfig;
@@ -55,44 +54,6 @@ namespace QuickStackStore
             return yPosCompare != 0 ? yPosCompare : a.x.CompareTo(b.x);
         }
 
-        private static bool hasCurrentlyToggledFavoriting = false;
-
-        private const string blackStar = "\u2605";
-        private const string whiteStar = "\u2606";
-
-        internal static bool HasCurrentlyToggledFavoriting
-        {
-            get => hasCurrentlyToggledFavoriting;
-            set
-            {
-                hasCurrentlyToggledFavoriting = value;
-
-                if (ButtonRenderer.favoritingTogglingButtonText != null)
-                {
-                    var color = ColorUtility.ToHtmlStringRGB(FavoriteConfig.BorderColorFavoritedItem.Value);
-
-                    ButtonRenderer.favoritingTogglingButtonText.text = $"<color=#{color}>{(value ? blackStar : whiteStar)}</color>";
-                }
-            }
-        }
-
-        internal static bool IsInFavoritingMode()
-        {
-            return HasCurrentlyToggledFavoriting
-                || Input.GetKey(FavoriteConfig.FavoritingModifierKey1.Value)
-                || Input.GetKey(FavoriteConfig.FavoritingModifierKey2.Value);
-        }
-
-        internal static Color GetMixedColor(Color color1, Color color2)
-        {
-            float r = (color1.r + color2.r) / 2f;
-            float g = (color1.g + color2.g) / 2f;
-            float b = (color1.b + color2.b) / 2f;
-            float a = (color1.a + color2.a) / 2f;
-
-            return new Color(r, g, b, a);
-        }
-
         // originally from 'Trash Items' mod, as allowed in their permission settings on nexus
         // https://www.nexusmods.com/valheim/mods/441
         // https://github.com/virtuaCode/valheim-mods/tree/main/TrashItems
@@ -109,62 +70,6 @@ namespace QuickStackStore
                 tex.LoadImage(mStream.ToArray());
                 tex.Apply();
                 return Sprite.Create(tex, size, pivot, units);
-            }
-        }
-
-        // taken from https://github.com/aedenthorn/ValheimMods AedenthornUtils.IgnoreKeyPresses, public domain
-        public static bool IgnoreKeyPresses()
-        {
-            // removed InventoryGui.IsVisible() because we specifically want that to be the case
-            return ZNetScene.instance == null || Player.m_localPlayer == null || Minimap.IsOpen() || Console.IsVisible() || TextInput.IsVisible() || ZNet.instance.InPasswordDialog() || Chat.instance?.HasFocus() == true || StoreGui.IsVisible() || Menu.IsVisible() || TextViewer.instance?.IsVisible() == true;
-        }
-    }
-
-    [HarmonyPatch(typeof(Player), nameof(Player.Update))]
-    public static class HotkeyListener
-    {
-        public static void Postfix(Player __instance)
-        {
-            if (GeneralConfig.OverrideKeybindBehavior.Value == OverrideKeybindBehavior.DisableAllNewHotkeys)
-            {
-                return;
-            }
-
-            if (Player.m_localPlayer != __instance)
-            {
-                return;
-            }
-
-            if (Helper.IgnoreKeyPresses())
-            {
-                return;
-            }
-
-            if (Input.GetKeyDown(QuickStackConfig.QuickStackKey.Value))
-            {
-                QuickStackRestockModule.DoQuickStack(__instance);
-            }
-            else if (Input.GetKeyDown(RestockConfig.RestockKey.Value))
-            {
-                QuickStackRestockModule.DoRestock(__instance);
-            }
-
-            if (!InventoryGui.IsVisible())
-            {
-                return;
-            }
-
-            if (Input.GetKeyDown(SortConfig.SortKey.Value))
-            {
-                SortModule.DoSort(__instance);
-            }
-            else if (Input.GetKeyDown(TrashConfig.TrashHotkey.Value))
-            {
-                TrashModule.TrashOrTrashFlagItem(true);
-            }
-            else if (Input.GetKeyDown(TrashConfig.QuickTrashHotkey.Value))
-            {
-                TrashModule.AttemptQuickTrash();
             }
         }
     }
