+++ QuickStackStore/Source/UI/ButtonRenderer.cs
@@ -49,7 +49,7 @@ namespace QuickStackStore
                 if (Player.m_localPlayer)
                 {
                     // reset in case player forgot to turn it off
-                    Helper.HasCurrentlyToggledFavoriting = false;
+                    FavoritingMode.HasCurrentlyToggledFavoriting = false;
 
                     var conf = SortConfig.AutoSort.Value;
 
@@ -254,12 +254,12 @@ namespace QuickStackStore
                         favoritingTogglingButtonText = favoritingTogglingButton.transform.Find("Text").GetComponent<Text>();
 
                         // trigger text reset without changing value
-                        Helper.HasCurrentlyToggledFavoriting |= false;
+                        FavoritingMode.HasCurrentlyToggledFavoriting |= false;
 
                         __instance.StartCoroutine(WaitAFrameToRepositionMiniButton(__instance, favoritingTogglingButton.transform, parent, index, randyStatus));
 
                         favoritingTogglingButton.onClick.RemoveAllListeners();
-                        favoritingTogglingButton.onClick.AddListener(new UnityAction(() => Helper.HasCurrentlyToggledFavoriting ^= true));
+                        favoritingTogglingButton.onClick.AddListener(new UnityAction(() => FavoritingMode.HasCurrentlyToggledFavoriting ^= true));
                     }
                     else
                     {
